Labelling confusion matrices properly
Contingent on pandas get_dummies returning an alphabetical order (which seems unlikely to change but ya never know i guess), as this knowledge is used a priori to map predictions to labels (via params.py). We then pass labels in a custom non-alphabetical order to the confusion matrix to get it in a nice layout.